pyproject.toml
[project]
name = "mlops-proyecto-final-equipo8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "mlflow>=3.11.1",
    "numpy>=2.4.4",
    "pandas>=2.3.3",
    "prefect>=3.6.27",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.44.0",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
